requirements.txt
requests
pillow
youtube-transcript-api
html2text
pydbus
matplotlib 